requirements.txt
opencv-python
torch
torchvision
numpy
pillow
tkinter
